internal-knowledge-navigator/agentthink_source/requirements.txt
# Core dependencies
python-dotenv
chromadb
google-generativeai
langchain
PyMuPDF
pdfplumber
tqdm
streamlit
flask
requests

# Optional utilities
pandas
numpy
